Fix type pun size ambiguity
sizeof(unsigned long) is not guaranteed to equal sizeof(int); while this
technically doesn't matter since the punned value is truncated to sizeof
the destination, this would break the sizeof equality assert in our
pun(...) helper func for such a case.

+++ src/public/dt_send.cpp
@@ -179,7 +179,7 @@ void SendProxy_UInt16ToInt32( const SendProp *pProp, const void *pStruct, const 
 void SendProxy_UInt32ToInt32( const SendProp *pProp, const void *pStruct, const void *pData, DVariant *pOut, int iElement, int objectID)
 {
 #ifdef NEO
-	pun<unsigned long>(pOut->m_Int, pData);
+	pun<uint32_t>(pOut->m_Int, pData);
 #else
 	*((unsigned long*)&pOut->m_Int) = *((unsigned long*)pData);
 #endif
